android: use v1.3 image for the base image
Within a build process of android images, we can no longer install a
right emulator which we can test arm64 image on x86_host.  This is
because the sdkmanager can install the latest (need confirmation?)
version of emulator, and the latest one as of today (2025 Feb.) doesn't
allow us to run arm64 image on x86_64 hosts.

Thus we use the previously built image, lkldocker/circleci-android:v1.3,
as a base image for android-arm{32,64} images.

We may need to overhaul the build/test infrastructure in future but this
is a quick fix which solves the issue we faced.

Signed-off-by: Hajime Tazaki <thehajime@gmail.com>